De-uint MIDI MT-32 and Fluidsynth
Threaded mixer can overflow the uint16_t in an edge case in fast-forward mode

+++ src/midi/midi_mt32.h
@@ -79,8 +79,8 @@ private:
 	void MixerCallBack(const int requested_audio_frames);
 	void ProcessWorkFromFifo();
 
-	uint16_t GetNumPendingAudioFrames();
-	void RenderAudioFramesToFifo(const uint16_t num_frames = 1);
+	int GetNumPendingAudioFrames();
+	void RenderAudioFramesToFifo(const int num_frames = 1);
 	void Render();
 
 	// Managed objects
